moved functions to correct extensions
+ removed internal symbols that should not be seen in PHP
