+++ src/Peachpie.Library.XmlDom/XsltProcessor.cs
@@ -21,7 +21,7 @@ namespace Peachpie.Library.XmlDom
         Always = -1
     }
 
-    [PhpExtension("dom")]
+    [PhpExtension("xsl")]
     public static class XsltConstants
     {
         public const int XSL_CLONE_AUTO = (int)CloneType.Auto;
